+++ packages/web/src/stores/user.ts
@@ -125,3 +125,4 @@ export const user = registerStore((manager) => {
   })
 
 })
+